feat: Refactor VRTs to use element-specific snapshots
This commit refactors the visual regression test suite to use a more granular, component-based approach. The `vrt.spec.ts` file has been updated to include tests for individual components, and the full-page `visual-regression.spec.ts` file has been removed.

While the new structure is in place, several tests were found to be flaky due to timeout issues. The following tests have been skipped and should be addressed in a separate ticket:

- `TimerDisplay Component`
- `MockHRM Client`
- `HR-Related Components`
